+++ java/src/main/java/gherkin/formatter/model/DataTableRow.java
@@ -6,11 +6,11 @@ public class DataTableRow extends Row {
     private transient final DiffType diffType;
 
 
-    public DataTableRow(List<Comment> comments, List<String> cells, int line) {
+    public DataTableRow(List<Comment> comments, List<String> cells, Integer line) {
         this(comments, cells, line, DiffType.NONE);
     }
 
-    public DataTableRow(List<Comment> comments, List<String> cells, int line, DiffType diffType) {
+    public DataTableRow(List<Comment> comments, List<String> cells, Integer line, DiffType diffType) {
         super(comments, cells, line);
         this.diffType = diffType;
     }