+++ spec/gherkin/i18n_spec.rb
@@ -165,7 +165,6 @@ module Gherkin
           Gherkin::I18n.keyword_regexp(:step).should =~ /\|Quando \|Quand \|Quan \|Pryd \|Pokud \|/
         end
 
-        unless defined?(JRUBY_VERSION)
         it "should print available languages" do
           ("\n" + Gherkin::I18n.language_table).should == %{
       | ar        | Arabic              | العربية           |
@@ -214,7 +213,6 @@ module Gherkin
       | zh-CN     | Chinese simplified  | 简体中文              |
       | zh-TW     | Chinese traditional | 繁體中文              |
 }
-        end
         end
 
         it "should print keywords for a given language" do