+++ History.md
@@ -1,3 +1,7 @@
+## [Git master](https://github.com/cucumber/gherkin/compare/v2.9.3...master)
+
+* [Java] Line numbers are Integer instead of int or Long (Aslak Hellesøy)
+
 ## [2.9.3](https://github.com/cucumber/gherkin/compare/v2.9.2...v2.9.3)
 
 * [Java] Fixed a bug in PrettyFormatter.setMonochrome(false) (Aslak Hellesøy)